+++ src/types/Categories.ts
@@ -1,0 +1,11 @@
+export interface Categories {
+    categories: Category[]
+  }
+  
+export interface Category {
+    idCategory: string
+    strCategory: string
+    strCategoryThumb: string
+    strCategoryDescription: string
+  }
+